# backend/openapi.yaml
openapi: 3.0.3
info:
  title: Accessible Games Platform API
  version: "1.0.0"
  description: >
    HTTP API for authentication, games catalog, user preferences, follows,
    reviews, and voice intent parsing used by the Team 13 project.
servers:
  - url: http://localhost:5000
    description: Local development
...
paths:
  /api/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [username, email, password]
              properties:
                username:
                  type: string
                  example: player1
                email:
                  type: string
                  format: email
                  example: player1@example.com
                password:
                  type: string
                  format: password
                  example: changeme123
      responses:
        "201":
          description: User created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing fields
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: Username or email already used
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Registration failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/login:
    post:
      tags: [Auth]
      summary: Log in with username or email
      description: Supply either `identifier`, `username`, or `email` plus `password`.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [password]
              properties:
                identifier:
                  type: string
                  description: Username or email
                  example: player1
                username:
                  type: string
                  example: player1
                email:
                  type: string
                  format: email
                  example: player1@example.com
                password:
                  type: string
                  format: password
                  example: changeme123
      responses:
        "200":
          description: Login succeeded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "400":
          description: Missing identifier/password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Invalid password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Login failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/auth/me:
    get:
      tags: [Auth]
      summary: Get the current user profile
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Authenticated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: User not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Load failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/games/search:
    get:
      tags: [Library]
      summary: Search games by text and tags
      parameters:
        - name: q
          in: query
          required: false
          description: Free-text search applied to title and platform.
          schema:
            type: string
          example: puzzle
        - name: tags
          in: query
          required: false
          description: Comma-separated list of tag names; all provided tags must match.
          schema:
            type: string
          example: Vision,Colourblind Mode
      responses:
        "200":
          description: Matching games
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/GameListItem"
        "500":
          description: Search failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/games/{id}:
    get:
      tags: [Games]
      summary: Get a single game with reviews
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Game found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GameDetail"
        "400":
          description: Invalid id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Game not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/games/{id}/reviews:
    get:
      tags: [Games]
      summary: List reviews for a game
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      responses:
        "200":
          description: Reviews loaded
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/ReviewWithUser"
        "500":
          description: Failed to fetch reviews
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    post:
      tags: [Games]
      summary: Create a review for a game
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
          example: 1
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [rating]
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  example: 5
                comment:
                  type: string
                  example: Loved the accessibility options.
      responses:
        "201":
          description: Review created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Review"
        "400":
          description: Missing rating
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Missing or invalid token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Failed to create review
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
        detail:
          type: string
    AuthResponse:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        token:
          type: string
          description: JWT bearer token
      example:
        id: 12
        username: player1
        email: player1@example.com
        token: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        username:
          type: string
        email:
          type: string
          format: email
        createdAt:
          type: string
          format: date-time
      example:
        id: 12
        username: player1
        email: player1@example.com
        createdAt: "2024-05-01T12:34:56.000Z"
...
    GameListItem:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        platform:
          type: string
        releaseDate:
          type: string
          nullable: true
          format: date
        rating:
          type: number
          format: float
        images:
          type: array
          items:
            type: string
            description: Path to thumbnail image
        tags:
          type: array
          items:
            type: string
      example:
        id: 1
        title: Cozy Quest
        platform: PC
        releaseDate: "2024-09-01"
        rating: 4.5
        images:
          - /images/cozy-quest-1.jpg
        tags: [Casual, Large Text, High Contrast]
    GameDetail:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        platform:
          type: string
        developer:
          type: string
        category:
          type: string
        releaseDate:
          type: string
          format: date
          nullable: true
        rating:
          type: number
        description:
          type: string
        images:
          type: array
          items:
            type: string
        tags:
          type: array
          items:
            type: object
            properties:
              id:
                type: integer
              name:
                type: string
        reviews:
          type: array
          items:
            $ref: "#/components/schemas/ReviewWithUser"
    Review:
      type: object
      properties:
        id:
          type: integer
        rating:
          type: integer
          minimum: 1
          maximum: 5
        comment:
          type: string
          nullable: true
        gameId:
          type: integer
        userId:
          type: integer
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ReviewWithUser:
      allOf:
        - $ref: "#/components/schemas/Review"
        - type: object
          properties:
            user:
              type: object
              nullable: true
              properties:
                id:
                  type: integer
                username:
                  type: string
                email:
                  type: string
                  format: email
                  nullable: true
...